+++ master/src/services/master/master-Role.service.ts
@@ -3,27 +3,34 @@ import {
   HttpStatus,
   Injectable,
   NotFoundException,
-} from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { MasterRoleDto } from 'src/dto/masterRole.dto';
-import { MasterRole } from 'src/entities/masterRole.entity';
-import { Repository } from 'typeorm';
+} from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { MasterRoleDto } from "src/dto/masterRole.dto";
+import { MasterRole } from "src/entities/masterRole.entity";
+import { Repository } from "typeorm";
 
 @Injectable()
 export class MasterRoleService {
   constructor(
     @InjectRepository(MasterRole)
-    private roleRepository: Repository<MasterRole>,
+    private roleRepository: Repository<MasterRole>
   ) {}
 
   public async create(masterRoleDto: MasterRoleDto) {
     try {
-      const newRole = await this.roleRepository.create(masterRoleDto);
-      if (newRole) {
-        await this.roleRepository.save(newRole);
-        return { data: newRole, message: 'Create Master-Role Successfully' };
+      const role = await this.roleRepository.findOne({
+        where: [{ name: masterRoleDto.name }],
+      });
+      if (role) {
+        const newRole = this.roleRepository.create(masterRoleDto);
+        if (newRole) {
+          await this.roleRepository.save(newRole);
+          return { data: newRole, message: "Create Master-Role Successfully" };
+        } else {
+          return { data: [], message: "fail" };
+        }
       } else {
-        return { data: [], message: 'fail' };
+        return { data: [], message: "Role Already Exists" };
       }
     } catch (err) {
       console.log(err);
@@ -35,9 +42,9 @@ export class MasterRoleService {
     try {
       const newRole = await this.roleRepository.find();
       if (newRole) {
-        return { data: newRole, message: 'Get All Role' };
+        return { data: newRole, message: "Get All Role" };
       } else {
-        return { data: [], message: 'Role Not Get' };
+        return { data: [], message: "Role Not Get" };
       }
     } catch (err) {
       console.log(err);
@@ -53,9 +60,9 @@ export class MasterRoleService {
         where: [{ id: id }],
       });
       if (newRole) {
-        return { data: newRole, message: 'Get Single Role' };
+        return { data: newRole, message: "Get Single Role" };
       } else {
-        return { data: [], message: 'Role Not Get' };
+        return { data: [], message: "Role Not Get" };
       }
     } catch (err) {
       console.log(err);
@@ -66,13 +73,21 @@ export class MasterRoleService {
 
   public async RoleDelete(id: number) {
     try {
-      // const newSkill = await skillRepository.findOne(id);
-      const newRole = await this.roleRepository.delete(id);
-      if (newRole) {
-        return { data: [], message: 'Delete Role' };
+      const role = await this.roleRepository.findOne({
+        where: [{ id: id }],
+      });
+      if (role) {
+        const newRole = await this.roleRepository.delete(id);
+        if (newRole) {
+          return { data: [], message: "Delete Role" };
+        } else {
+          return { data: [], message: "Role Not Delete" };
+        }
       } else {
-        return { data: [], message: 'Role Not Delete' };
+        return { data: [], message: "Role Not Exists" };
       }
+
+      // const newSkill = await skillRepository.findOne(id);
     } catch (err) {
       console.log(err);
       return err;
@@ -82,14 +97,21 @@ export class MasterRoleService {
 
   public async RoleUpdate(id: number, masterRoleDto: MasterRoleDto) {
     try {
-      // const newSkill = await skillRepository.findOne(id);
-      const updateRole = await this.roleRepository.update(id, masterRoleDto);
+      const role = await this.roleRepository.findOne({
+        where: [{ id: id }],
+      });
+      if (role) {
+        const updateRole = await this.roleRepository.update(id, masterRoleDto);
 
-      if (updateRole) {
-        return { data: updateRole, message: 'Update Role' };
+        if (updateRole) {
+          return { data: updateRole, message: "Update Role" };
+        } else {
+          return { data: [], message: "Role Not Update" };
+        }
       } else {
-        return { data: [], message: 'Role Not Update' };
+        return { data: [], message: "Role Not Exists" };
       }
+      // const newSkill = await skillRepository.findOne(id);
     } catch (err) {
       console.log(err);
       return err;
