+++ master/src/services/master/master-Skill.service.ts
@@ -25,9 +25,16 @@ export class masterSkillService {
 
   public async create(masterSkillDTO: MasterSkillDto) {
     try {
-      const newSkill = await this.skillRepository.create(masterSkillDTO);
-      await this.skillRepository.save(newSkill);
-      return newSkill;
+      const skill = await this.skillRepository.findOne({
+        where: [{ name: masterSkillDTO.name }],
+      });
+      if (skill) {
+        const newSkill = await this.skillRepository.create(masterSkillDTO);
+        await this.skillRepository.save(newSkill);
+        return { data: newSkill, message: "Skill Add Successfully" };
+      } else {
+        return { data: [], message: "Skill Already Exists" };
+      }
     } catch (err) {
       console.log(err);
       return err;
@@ -52,7 +59,6 @@ export class masterSkillService {
 
   public async getSkillById(id: number) {
     try {
-      // const newSkill = await skillRepository.findOne(id);
       const newSkill = await this.skillRepository.findOne({
         where: [{ id: id }],
       });
@@ -61,6 +67,8 @@ export class masterSkillService {
       } else {
         return { data: [], message: "Skill Not Get" };
       }
+
+      // const newSkill = await skillRepository.findOne(id);
     } catch (err) {
       console.log(err);
       return err;
@@ -70,13 +78,20 @@ export class masterSkillService {
 
   public async skillDelete(id: number) {
     try {
-      // const newSkill = await skillRepository.findOne(id);
-      const newSkill = await this.skillRepository.delete(id);
-      if (newSkill) {
-        return { data: [], message: "Delete Skill" };
+      const skill = await this.skillRepository.findOne({
+        where: [{ id: id }],
+      });
+      if (skill) {
+        const newSkill = await this.skillRepository.delete(id);
+        if (newSkill) {
+          return { data: [], message: "Delete Skill" };
+        } else {
+          return { data: [], message: "Skill Not Delete" };
+        }
       } else {
-        return { data: [], message: "Skill Not Delete" };
+        return { data: [], message: "Company Not Exists" };
       }
+      // const newSkill = await skillRepository.findOne(id);
     } catch (err) {
       console.log(err);
       return err;
@@ -86,17 +101,27 @@ export class masterSkillService {
 
   public async skillUpdate(id: number, masterSkillDTO: MasterSkillDto) {
     try {
-      // const newSkill = await skillRepository.findOne(id);
-      const updateSkill = await this.skillRepository.update(id, masterSkillDTO);
+      const skill = await this.skillRepository.findOne({
+        where: [{ id: id }],
+      });
+      if (skill) {
+        const updateSkill = await this.skillRepository.update(
+          id,
+          masterSkillDTO
+        );
 
-      if (updateSkill) {
-        const newSkill = await this.skillRepository.findOne({
-          where: [{ id: id }],
-        });
-        return { data: newSkill, message: "Update Skill" };
+        if (updateSkill) {
+          const newSkill = await this.skillRepository.findOne({
+            where: [{ id: id }],
+          });
+          return { data: newSkill, message: "Update Skill" };
+        } else {
+          return { data: [], message: "Skill Not Update" };
+        }
       } else {
-        return { data: [], message: "Skill Not Update" };
+        return { data: [], message: "Company Not Exists" };
       }
+      // const newSkill = await skillRepository.findOne(id);
     } catch (err) {
       console.log(err);
       return err;
@@ -105,47 +130,6 @@ export class masterSkillService {
   }
 
   public async payment(res) {
-    // const request = require("request");
-
-    // const orderId = "Order001";
-    // const orderAmount = "100.00";
-    // const orderCurrency = "INR";
-    // const customerName = "John Doe";
-    // const customerEmail = "johndoe@example.com";
-    // const customerPhone = "9999999999";
-    // const returnUrl = "http://example.com/success";
-    // const notifyUrl = "http://example.com/notify";
-
-    // const options = {
-    //   url: "https://test.cashfree.com/api/v2/cftoken/order",
-    //   method: "POST",
-    //   headers: {
-    //     "Content-Type": "application/json",
-    //     "x-client-id": "TEST377598107ac66ee0a5f1faae0f895773",
-    //     "x-client-secret": "TEST1cd93ad8a2cd10329d44017a731a5fbc629124c4",
-    //   },
-    //   json: true,
-    //   body: {
-    //     orderId,
-    //     orderAmount,
-    //     orderCurrency,
-    //     customerName,
-    //     customerEmail,
-    //     customerPhone,
-    //     returnUrl,
-    //     notifyUrl,
-    //   },
-    // };
-
-    // request(options, function (error, response, body) {
-    //   if (error) {
-    //     console.error(error);
-    //   } else {
-    //     console.log(response);
-
-    //     return response;
-    //   }
-    // });
     const request = require("request");
 
     const options = {
@@ -175,8 +159,7 @@ export class masterSkillService {
       if (error) {
         console.error(error);
       } else {
-        console.log(body);
-        return { data: body, message: "sucess" };
+        res.json({ data: body, message: "Order Created" });
       }
     });
   }
